Add app factory test
